refactor: ThemeStyleSet 타입 추가
